Format, prepare for MR
